文章排序

+++ article/views.py
@@ -14,12 +14,39 @@ from .forms import ArticlePostForm
 # 引入User模型
 from django.contrib.auth.models import User
 
+from django.contrib.auth.decorators import login_required
+
+# 引入分页模块
+from django.core.paginator import Paginator
+
+# 引入 Q 对象
+from django.db.models import Q
+
+#from django.views.decorators.csrf import csrf_exempt
+
 # 视图函数
+#@csrf_exempt
 def article_list(request):
     # 取出所有博客文章
-    articles = ArticlePost.objects.all()
+    # 分页处理 修改变量名称（articles -> article_list）
+    #article_list = ArticlePost.objects.all()
+
+    # 根据GET请求中查询条件
+    # 返回不同排序的对象数组
+    if request.GET.get('order') == 'total_views':
+        article_list = ArticlePost.objects.all().order_by('-total_views')
+        order = 'total_views'
+    else:
+        article_list = ArticlePost.objects.all()
+        order = 'normal'
+    # 每页显示 6 篇文章
+    paginator = Paginator(article_list, 6)
+    # 获取 url 中的页码
+    page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
     # 需要传递给模板（templates）的对象
-    context = { 'articles': articles }
+    context = { 'articles': articles,'order': order,'search': search }
     # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
@@ -31,6 +58,10 @@ import markdown
 def article_detail(request, id):
     # 取出相应的文章
     article = ArticlePost.objects.get(id=id)
+
+    # 浏览量 +1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
 
     # 将markdown语法渲染成html样式
     article.body = markdown.markdown(article.body,
@@ -46,7 +77,8 @@ def article_detail(request, id):
     # 载入模板，并返回context对象
     return render(request, 'article/detail.html', context)
 
-
+# 检查登录
+@login_required(login_url='/userprofile/login/')
 # 写文章的视图
 def article_create(request):
     # 判断用户是否提交数据
@@ -60,7 +92,7 @@ def article_create(request):
             # 指定数据库中 id=1 的用户为作者
             # 如果你进行过删除数据表的操作，可能会找不到id=1的用户
             # 此时请重新创建用户，并传入此用户的id
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存到数据库中
             new_article.save()
             # 完成后返回到文章列表
@@ -77,21 +109,67 @@ def article_create(request):
         # 返回模板
         return render(request, 'article/create.html', context)
 
-
+#@login_required(login_url='/userprofile/login/')
 # 删文章
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    #if request.user != article.author:
+        #return HttpResponse("抱歉，你无权删除这篇文章。")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
     return redirect("article:article_list")
 
+@login_required(login_url='/userprofile/login/')
 # 安全删除文章
 def article_safe_delete(request, id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse("抱歉，你无权删除这篇文章。")
         article.delete()
         return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
+
+# 提醒用户登录
+@login_required(login_url='/userprofile/login/')
+# 更新文章
+def article_update(request, id):
+    """
+    更新文章的视图函数
+    通过POST方法提交表单，更新titile、body字段
+    GET方法进入初始表单页面
+    id： 文章的 id
+    """
+
+    # 获取需要修改的具体文章对象
+    article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
+    # 判断用户是否为 POST 提交表单数据
+    if request.method == "POST":
+        # 将提交的数据赋值到表单实例中
+        article_post_form = ArticlePostForm(data=request.POST)
+        # 判断提交的数据是否满足模型的要求
+        if article_post_form.is_valid():
+            # 保存新写入的 title、body 数据并保存
+            article.title = request.POST['title']
+            article.body = request.POST['body']
+            article.save()
+            # 完成后返回到修改后的文章中。需传入文章的 id 值
+            return redirect("article:article_detail", id=id)
+        # 如果数据不合法，返回错误信息
+        else:
+            return HttpResponse("表单内容有误，请重新填写。")
+
+    # 如果用户 GET 请求获取数据
+    else:
+        # 创建表单类实例
+        article_post_form = ArticlePostForm()
+        # 赋值上下文，将 article 文章对象也传递进去，以便提取旧的内容
+        context = { 'article': article, 'article_post_form': article_post_form }
+        # 将响应返回到模板中
+        return render(request, 'article/update.html', context)
